requirements-cpu.txt
#./requirements-cpu.txt

# Core
numpy==1.23.5
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.10.1
tqdm==4.65.0
python-dotenv==1.0.0

# Audio Processing
librosa==0.10.0
soundfile==0.12.1

# PyTorch (CPU-only)
torch==2.0.1+cpu --index-url https://download.pytorch.org/whl/cpu
torchaudio==2.0.2+cpu --index-url https://download.pytorch.org/whl/cpu
torchvision==0.15.2+cpu --index-url https://download.pytorch.org/whl/cpu

# Deep Learning
pytorch-lightning==2.0.0
transformers==4.30.0

# Evaluation & Logging
matplotlib==3.7.2
seaborn==0.12.2
wandb==0.15.0

# Data Management
huggingface-hub==0.16.0
h5py==3.9.0

# CPU Optimization
onnxruntime==1.15.1